+++ R/LearnerClassifXgboost.R
@@ -105,9 +105,7 @@ LearnerClassifXgboost = R6Class("LearnerClassifXgboost",
         save_name                   = p_uty(default = NULL, tags = "train"),
         save_period                 = p_int(0, default = NULL, special_vals = list(NULL), tags = "train"),
         scale_pos_weight            = p_dbl(default = 1, tags = "train"),
-        sketch_eps                  = p_dbl(0, 1, default = 0.03, tags = "train"),
         skip_drop                   = p_dbl(0, 1, default = 0, tags = "train"),
-        single_precision_histogram  = p_lgl(default = FALSE, tags = "train"),
         strict_shape                = p_lgl(default = FALSE, tags = "predict"),
         subsample                   = p_dbl(0, 1, default = 1, tags = c("train", "control")),
         top_k                       = p_int(0, default = 0, tags = "train"),
@@ -133,11 +131,9 @@ LearnerClassifXgboost = R6Class("LearnerClassifXgboost",
       ps$add_dep("grow_policy", "tree_method", CondEqual$new("hist"))
       ps$add_dep("max_leaves", "grow_policy", CondEqual$new("lossguide"))
       ps$add_dep("max_bin", "tree_method", CondEqual$new("hist"))
-      ps$add_dep("sketch_eps", "tree_method", CondEqual$new("approx"))
       ps$add_dep("feature_selector", "booster", CondEqual$new("gblinear"))
       ps$add_dep("top_k", "booster", CondEqual$new("gblinear"))
       ps$add_dep("top_k", "feature_selector", CondAnyOf$new(c("greedy", "thrifty")))
-      ps$add_dep("single_precision_histogram", "tree_method", CondEqual$new("hist"))
       ps$add_dep("lambda_bias", "booster", CondEqual$new("gblinear"))
 
       # custom defaults
@@ -164,7 +160,6 @@ LearnerClassifXgboost = R6Class("LearnerClassifXgboost",
       }
 
       imp = xgboost::xgb.importance(
-        feature_names = self$model$features,
         model = self$model
       )
       set_names(imp$Gain, imp$Feature)